Moved namespace because Rider was unhappy
