+++ BTCPayServer.Abstractions/Custodians/Client/MarketTradeResult.cs
@@ -1,7 +1,7 @@
 using System.Collections.Generic;
 using BTCPayServer.Client.Models;
 
-namespace BTCPayServer.Abstractions.Custodians;
+namespace BTCPayServer.Abstractions.Custodians.Client;
 
 /**
  * The result of a market trade. Used as a return type for custodians implementing ICanTrade
