+++ BTCPayServer.Abstractions/Custodians/ICanTrade.cs
@@ -1,6 +1,7 @@
 using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
+using BTCPayServer.Abstractions.Custodians.Client;
 using BTCPayServer.Client.Models;
 using Newtonsoft.Json.Linq;
