+++ BTCPayServer/Models/CustodianAccountViewModels/TradePrepareViewModel.cs
@@ -1,3 +1,5 @@
+using BTCPayServer.Abstractions.Custodians.Client;
+
 namespace BTCPayServer.Models.CustodianAccountViewModels;
 
 public class TradePrepareViewModel
